backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-okx>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiosqlite>=0.21.0
httpx>=0.28.0
pydantic>=2.0.0
